Rename to base and add await to run example in docs
