+++ src/neo4j_graphrag/embeddings/base.py
