+++ tests/unit/conftest.py
@@ -18,7 +18,7 @@ from unittest.mock import MagicMock, patch
 
 import neo4j
 import pytest
-from neo4j_graphrag.embeddings.embedder import Embedder
+from neo4j_graphrag.embeddings.base import Embedder
 from neo4j_graphrag.llm import LLMInterface
 from neo4j_graphrag.retrievers import (
     HybridRetriever,
